bioai-analyzer-backend/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
biopython==1.81
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uvicorn[standard]==0.24.0
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
